+++ packages/components/src/Modal/Layout/ModalFooter.test.tsx
@@ -47,7 +47,7 @@ test('ModalFooter with ModalContext', () => {
           <button
             onClick={() => {
               alert("doin' things...")
-              close && close()
+              closeModal && closeModal()
             }}
           >
             Yes, Delete "Stuff"