+++ packages/components/src/utils/index.ts
@@ -26,5 +26,7 @@
 
 export * from './GlobalStyle'
 export * from './useControlWarn'
+export * from './useCallbackRef'
 export * from './useFocusTrap'
+export * from './useScrollLock'
 export * from './useToggle'